refactor(trill): moved terminal resize handling into platform-specific functions and added support for Windows
Moved the initial pseudo-TTY resize to *after* the container is started, as Docker becomes upset when you try to do it while the container is still stopped.

Terminal resizing changes on Windows relies on a polling-based model. I have no idea what the performance implications of this is, if any, but it seems to work well enough.